+++ app/src/main/java/com/oneflow/analytics/utils/OFMyCountDownTimer.java
@@ -71,48 +71,51 @@ public class OFMyCountDownTimer extends CountDownTimer implements OFMyResponseHa
 
                 OFHelper.v("FeedbackController", "OneFlow fetchEventsFromDB came back");
 
+                if(obj!=null) {
+                    ArrayList<OFRecordEventsTab> list = (ArrayList<OFRecordEventsTab>) obj;
+                    OFHelper.v("FeedbackController", "OneFlow fetchEventsFromDB list received size[" + list.size() + "]");
+                    //Preparing list to send api
+                    if (list.size() > 0) {
+                        Integer[] ids = new Integer[list.size()];
+                        int i = 0;
+                        ArrayList<OFRecordEventsTabToAPI> retListToAPI = new ArrayList<>();
+                        OFRecordEventsTabToAPI retMain;
+                        for (OFRecordEventsTab ret : list) {
+                            retMain = new OFRecordEventsTabToAPI();
+                            retMain.setPlatform("a");
+                            retMain.setEventName(ret.getEventName());
+                            retMain.setTime(ret.getTime());
+                            retMain.setDataMap(ret.getDataMap());
+                            retListToAPI.add(retMain);
 
-                ArrayList<OFRecordEventsTab> list = (ArrayList<OFRecordEventsTab>) obj;
-                OFHelper.v("FeedbackController", "OneFlow fetchEventsFromDB list received size[" + list.size() + "]");
-                //Preparing list to send api
-                if (list.size() > 0) {
-                    Integer[] ids = new Integer[list.size()];
-                    int i = 0;
-                    ArrayList<OFRecordEventsTabToAPI> retListToAPI = new ArrayList<>();
-                    OFRecordEventsTabToAPI retMain;
-                    for (OFRecordEventsTab ret : list) {
-                        retMain = new OFRecordEventsTabToAPI();
-                        retMain.setPlatform("a");
-                        retMain.setEventName(ret.getEventName());
-                        retMain.setTime(ret.getTime());
-                        retMain.setDataMap(ret.getDataMap());
-                        retListToAPI.add(retMain);
+                            ids[i++] = ret.getId();
+                        }
 
-                        ids[i++] = ret.getId();
+                        if (!new OFOneFlowSHP(mContext).getStringValue(OFConstants.SESSIONDETAIL_IDSHP).equalsIgnoreCase("NA")) {
+                            OFEventAPIRequest ear = new OFEventAPIRequest();
+                            ear.setSessionId(new OFOneFlowSHP(mContext).getStringValue(OFConstants.SESSIONDETAIL_IDSHP));
+                            ear.setEvents(retListToAPI);
+                            OFHelper.v("FeedbackController", "OneFlow fetchEventsFromDB request prepared");
+                            OFEventAPIRepo.sendLogsToApi(new OFOneFlowSHP(mContext).getStringValue(OFConstants.APPIDSHP), ear, this, OFConstants.ApiHitType.sendEventsToAPI, ids);
+                        }
+
                     }
-
-                    if (!new OFOneFlowSHP(mContext).getStringValue(OFConstants.SESSIONDETAIL_IDSHP).equalsIgnoreCase("NA")) {
-                        OFEventAPIRequest ear = new OFEventAPIRequest();
-                        ear.setSessionId(new OFOneFlowSHP(mContext).getStringValue(OFConstants.SESSIONDETAIL_IDSHP));
-                        ear.setEvents(retListToAPI);
-                        OFHelper.v("FeedbackController", "OneFlow fetchEventsFromDB request prepared");
-                        OFEventAPIRepo.sendLogsToApi(new OFOneFlowSHP(mContext).getStringValue(OFConstants.APPIDSHP), ear, this, OFConstants.ApiHitType.sendEventsToAPI, ids);
-                    }
-
                 }
                 break;
             case sendEventsToAPI:
-                //Events has been sent to api not deleting local records
-                Integer[] ids1 = (Integer[]) obj;
-                OFEventDBRepo.deleteEvents(mContext, ids1, this, OFConstants.ApiHitType.deleteEventsFromDB);
-
+                if(obj!=null) {
+                    //Events has been sent to api not deleting local records
+                    Integer[] ids1 = (Integer[]) obj;
+                    OFEventDBRepo.deleteEvents(mContext, ids1, this, OFConstants.ApiHitType.deleteEventsFromDB);
+                }
                 break;
             case deleteEventsFromDB:
-                OFHelper.v("FeedbackControler", "OneFlow events delete count[" + ((Integer) obj) + "]");
-                Intent intent = new Intent("events_submitted");
-                intent.putExtra("size", String.valueOf((Integer) obj));
-                mContext.sendBroadcast(intent);
-
+                if(obj!=null) {
+                    OFHelper.v("FeedbackControler", "OneFlow events delete count[" + ((Integer) obj) + "]");
+                    Intent intent = new Intent("events_submitted");
+                    intent.putExtra("size", String.valueOf((Integer) obj));
+                    mContext.sendBroadcast(intent);
+                }
                 break;
         }
     }
